move utils out of utils folder
